separate ResultMap to new file
