+++ resultmap.go
@@ -1,0 +1,50 @@
+package reviewdog
+
+import (
+	"errors"
+	"fmt"
+	"sync"
+)
+
+// ResultMap represents a concurrent-safe map to store CheckResults generated by concurrent jobs.
+type ResultMap struct {
+	sm sync.Map
+}
+
+// Store saves a new []*CheckResult into ResultMap.
+func (rm *ResultMap) Store(key string, crs []*CheckResult) {
+	rm.sm.Store(key, crs)
+}
+
+// Load fetches []*CheckResult from ResultMap
+func (rm *ResultMap) Load(key string) ([]*CheckResult, error) {
+	v, ok := rm.sm.Load(key)
+	if !ok {
+		return nil, fmt.Errorf("fail to get the value of key %q from results", key)
+	}
+
+	t, ok := v.([]*CheckResult)
+	if !ok {
+		return nil, errors.New("stored type in ResultMap is invalid")
+	}
+
+	return t, nil
+}
+
+// Range retrieves `key` and `values` from ResultMap iteratively.
+func (rm *ResultMap) Range(f func(key string, val []*CheckResult)) {
+	rm.sm.Range(func(k, v interface{}) bool {
+		f(k.(string), v.([]*CheckResult))
+		return true
+	})
+}
+
+// Len returns the length of ResultMap count. Len() is not yet officially not supported by Go. (ref: https://github.com/golang/go/issues/20680)
+func (rm *ResultMap) Len() int {
+	l := 0
+	rm.sm.Range(func(_, _ interface{}) bool {
+		l++
+		return true
+	})
+	return l
+}
