+++ reviewdog.go
@@ -3,11 +3,9 @@ package reviewdog
 import (
 	"bytes"
 	"context"
-	"errors"
 	"fmt"
 	"io"
 	"os"
-	"sync"
 
 	"github.com/haya14busa/reviewdog/diff"
 )
@@ -43,49 +41,6 @@ type CheckResult struct {
 	Col     int      // column number (1 <tab> == 1 character column)
 	Message string   // error message
 	Lines   []string // Original error lines (often one line)
-}
-
-// ResultMap represents a concurrent-safe map to store CheckResults generated by concurrent jobs.
-type ResultMap struct {
-	sm sync.Map
-}
-
-// Store saves a new []*CheckResult into ResultMap.
-func (rm *ResultMap) Store(key string, crs []*CheckResult) {
-	rm.sm.Store(key, crs)
-}
-
-// Load fetches []*CheckResult from ResultMap
-func (rm *ResultMap) Load(key string) ([]*CheckResult, error) {
-	v, ok := rm.sm.Load(key)
-	if !ok {
-		return nil, fmt.Errorf("fail to get the value of key %s from results", key)
-	}
-
-	t, ok := v.([]*CheckResult)
-	if !ok {
-		return nil, errors.New("stored type in ResultMap is invalid")
-	}
-
-	return t, nil
-}
-
-// Range retrieves `key` and `values` from ResultMap iteratively.
-func (rm *ResultMap) Range(f func(key string, val []*CheckResult)) {
-	rm.sm.Range(func(k, v interface{}) bool {
-		f(k.(string), v.([]*CheckResult))
-		return true
-	})
-}
-
-// Len returns the length of ResultMap count. Len() is not yet officially not supported by Go. (ref: https://github.com/golang/go/issues/20680)
-func (rm *ResultMap) Len() int {
-	l := 0
-	rm.sm.Range(func(_, _ interface{}) bool {
-		l++
-		return true
-	})
-	return l
 }
 
 // Parser is an interface which parses compilers, linters, or any tools
